backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
joblib==1.4.2
numpy==1.26.4
scikit-learn==1.4.2
textblob==0.18.0
openai-whisper==20231117
torch>=2.0.0
pandas==2.2.2